+++ vulyk/control.py
@@ -23,7 +23,7 @@ def cli():
     """Vulyk UA management CLI"""
     pass
 
-
+# region Admin
 @cli.group('admin')
 def admin():
     """Manages admin users"""
@@ -48,8 +48,9 @@ def admin_add(email):
 def admin_remove(email):
     """Unmark user as admin"""
     _admin.toggle_admin(email, False)
+# endregion Admin
 
-
+# region DB (export/import)
 @cli.group('db')
 def db():
     """Commands to manage DB"""
@@ -90,8 +91,9 @@ def load(task_type, path, batch):
 def export(task_type, path, batch, export_all):
     """Exports answers on closed tasks to json."""
     _db.export_tasks(TASKS_TYPES[task_type], path, batch, not export_all)
+# endregion DB (export/import)
 
-
+# region Group
 @cli.group('group')
 def group():
     """Groups management section"""
@@ -165,8 +167,9 @@ def group_addtype(gid, task_type):
               prompt='Provide the task type name')
 def group_deltype(gid, task_type):
     _groups.remove_task_type(gid, task_type=task_type)
+# endregion Group
 
-
+# region Bootstrapping
 @cli.command('init')
 @click.argument('allowed_types',
                 type=click.Choice(TASKS_TYPES.keys()),
@@ -182,8 +185,9 @@ def project_init(allowed_types):
                                  'one default task type')
 
     _project_init(allowed_types)
+# endregion Bootstrapping
 
-
+# region Stats
 @cli.group('stats')
 def stats():
     """Commands to show some stats"""
@@ -222,3 +226,4 @@ def batch(batch_name, task_type):
         pt.add_row(values)
 
     print(pt)
+# endregion Stats